perf: lazy load public auth shell
