+++ src/app/[locale]/advertise/new/page.tsx
@@ -6,6 +6,7 @@ import { getTranslations } from "next-intl/server";
 import { buildLocaleAlternates, withLocale } from "@/lib/locale-routing";
 
 import { AdvertiseForm } from "@/components/ads/advertise-form";
+import { FullAuthProviders } from "@/components/auth-providers";
 import { SiteFooter } from "@/components/site-footer";
 import { SiteHeader } from "@/components/site-header";
 
@@ -53,55 +54,57 @@ export default async function NewAdvertisePage({
   }
 
   return (
-    <main className="min-h-dvh bg-background text-foreground">
-      <SiteHeader />
-      <section className="petdex-cloud relative -mt-[84px] pt-[84px]">
-        <div className="relative mx-auto flex w-full max-w-[1440px] flex-col gap-8 px-5 pt-10 pb-16 md:px-8 md:pt-14 md:pb-20">
-          <header className="flex flex-col gap-5 md:flex-row md:items-end md:justify-between">
-            <div>
-              <p className="font-mono text-xs tracking-[0.22em] text-brand uppercase">
-                {t("create.eyebrow")}
-              </p>
-              <h1 className="mt-3 max-w-4xl text-balance text-4xl font-semibold tracking-tight md:text-6xl">
-                {t("create.title")}
-              </h1>
-              <p className="mt-4 max-w-2xl text-base leading-7 text-muted-2">
-                {t("create.body")}
-              </p>
-            </div>
-            <Link
-              href={withLocale("/advertise", localeValue)}
-              className="inline-flex h-11 shrink-0 items-center justify-center rounded-full border border-border-base bg-surface/72 px-5 text-sm font-medium text-foreground backdrop-blur transition hover:border-border-strong"
-            >
-              {t("create.backToLanding")}
-            </Link>
-          </header>
+    <FullAuthProviders>
+      <main className="min-h-dvh bg-background text-foreground">
+        <SiteHeader />
+        <section className="petdex-cloud relative -mt-[84px] pt-[84px]">
+          <div className="relative mx-auto flex w-full max-w-[1440px] flex-col gap-8 px-5 pt-10 pb-16 md:px-8 md:pt-14 md:pb-20">
+            <header className="flex flex-col gap-5 md:flex-row md:items-end md:justify-between">
+              <div>
+                <p className="font-mono text-xs tracking-[0.22em] text-brand uppercase">
+                  {t("create.eyebrow")}
+                </p>
+                <h1 className="mt-3 max-w-4xl text-balance text-4xl font-semibold tracking-tight md:text-6xl">
+                  {t("create.title")}
+                </h1>
+                <p className="mt-4 max-w-2xl text-base leading-7 text-muted-2">
+                  {t("create.body")}
+                </p>
+              </div>
+              <Link
+                href={withLocale("/advertise", localeValue)}
+                className="inline-flex h-11 shrink-0 items-center justify-center rounded-full border border-border-base bg-surface/72 px-5 text-sm font-medium text-foreground backdrop-blur transition hover:border-border-strong"
+              >
+                {t("create.backToLanding")}
+              </Link>
+            </header>
 
-          {checkout === "cancelled" ? (
-            <div className="rounded-2xl bg-chip-warning-bg p-4 text-sm leading-6 text-chip-warning-fg">
-              <p className="font-semibold">{t("checkout.cancelTitle")}</p>
-              <p className="mt-1">{t("checkout.cancelBody")}</p>
-            </div>
-          ) : null}
+            {checkout === "cancelled" ? (
+              <div className="rounded-2xl bg-chip-warning-bg p-4 text-sm leading-6 text-chip-warning-fg">
+                <p className="font-semibold">{t("checkout.cancelTitle")}</p>
+                <p className="mt-1">{t("checkout.cancelBody")}</p>
+              </div>
+            ) : null}
 
-          <div className="grid gap-5 lg:grid-cols-[1fr_280px] lg:items-start">
-            <AdvertiseForm />
-            <aside className="grid gap-3 lg:sticky lg:top-24">
-              <HelperCard title={t("create.help.creative.title")}>
-                {t("create.help.creative.body")}
-              </HelperCard>
-              <HelperCard title={t("create.help.review.title")}>
-                {t("create.help.review.body")}
-              </HelperCard>
-              <HelperCard title={t("create.help.billing.title")}>
-                {t("create.help.billing.body")}
-              </HelperCard>
-            </aside>
+            <div className="grid gap-5 lg:grid-cols-[1fr_280px] lg:items-start">
+              <AdvertiseForm />
+              <aside className="grid gap-3 lg:sticky lg:top-24">
+                <HelperCard title={t("create.help.creative.title")}>
+                  {t("create.help.creative.body")}
+                </HelperCard>
+                <HelperCard title={t("create.help.review.title")}>
+                  {t("create.help.review.body")}
+                </HelperCard>
+                <HelperCard title={t("create.help.billing.title")}>
+                  {t("create.help.billing.body")}
+                </HelperCard>
+              </aside>
+            </div>
           </div>
-        </div>
-      </section>
-      <SiteFooter />
-    </main>
+        </section>
+        <SiteFooter />
+      </main>
+    </FullAuthProviders>
   );
 }
