+++ src/app/[locale]/my-feedback/page.tsx
@@ -8,6 +8,7 @@ import { getTranslations } from "next-intl/server";
 
 import { db, schema } from "@/lib/db/client";
 
+import { FullAuthProviders } from "@/components/auth-providers";
 import { MyFeedbackFilters } from "@/components/my-feedback-filters";
 import { SiteFooter } from "@/components/site-footer";
 import { SiteHeader } from "@/components/site-header";
@@ -177,119 +178,124 @@ export default async function MyFeedbackPage({
   });
 
   return (
-    <main className="min-h-dvh bg-background text-foreground">
-      <SiteHeader />
-      <section className="mx-auto flex w-full max-w-3xl flex-col gap-6 px-5 pt-8 pb-20 md:px-8">
-        <header className="space-y-3">
-          <p className="font-mono text-xs tracking-[0.22em] text-brand uppercase">
-            {t("eyebrow")}
-          </p>
-          <h1 className="text-4xl font-medium tracking-tight md:text-5xl">
-            {t("title")}
-          </h1>
-          <p className="text-sm text-muted-2">{t("subtitle")}</p>
-          {decorated.length > 0 ? (
-            <MyFeedbackFilters counts={counts} defaultFilter={defaultFilter} />
-          ) : null}
-        </header>
+    <FullAuthProviders>
+      <main className="min-h-dvh bg-background text-foreground">
+        <SiteHeader />
+        <section className="mx-auto flex w-full max-w-3xl flex-col gap-6 px-5 pt-8 pb-20 md:px-8">
+          <header className="space-y-3">
+            <p className="font-mono text-xs tracking-[0.22em] text-brand uppercase">
+              {t("eyebrow")}
+            </p>
+            <h1 className="text-4xl font-medium tracking-tight md:text-5xl">
+              {t("title")}
+            </h1>
+            <p className="text-sm text-muted-2">{t("subtitle")}</p>
+            {decorated.length > 0 ? (
+              <MyFeedbackFilters
+                counts={counts}
+                defaultFilter={defaultFilter}
+              />
+            ) : null}
+          </header>
 
-        {decorated.length === 0 ? (
-          <div className="rounded-2xl border border-dashed border-border-base bg-surface/60 p-10 text-center text-sm text-stone-600 dark:text-stone-400">
-            {t("empty")}
-          </div>
-        ) : visible.length === 0 ? (
-          <div className="rounded-2xl border border-dashed border-border-base bg-surface/60 p-10 text-center text-sm text-muted-2">
-            {t("emptyFiltered")}
-          </div>
-        ) : (
-          <ul className="space-y-2">
-            {visible.map(({ row: r, agg, unread, replied, waiting }) => {
-              const meta = KIND_META[r.kind] ?? KIND_META.other;
-              const lastAdminAt = agg?.latestAdminAt ?? null;
-              const lastAdminBody = agg?.latestAdminBody ?? null;
-              const replyCount = agg?.replyCount ?? 0;
+          {decorated.length === 0 ? (
+            <div className="rounded-2xl border border-dashed border-border-base bg-surface/60 p-10 text-center text-sm text-stone-600 dark:text-stone-400">
+              {t("empty")}
+            </div>
+          ) : visible.length === 0 ? (
+            <div className="rounded-2xl border border-dashed border-border-base bg-surface/60 p-10 text-center text-sm text-muted-2">
+              {t("emptyFiltered")}
+            </div>
+          ) : (
+            <ul className="space-y-2">
+              {visible.map(({ row: r, agg, unread, replied, waiting }) => {
+                const meta = KIND_META[r.kind] ?? KIND_META.other;
+                const lastAdminAt = agg?.latestAdminAt ?? null;
+                const lastAdminBody = agg?.latestAdminBody ?? null;
+                const replyCount = agg?.replyCount ?? 0;
 
-              return (
-                <li key={r.id}>
-                  <Link
-                    href={`/my-feedback/${r.id}`}
-                    className={`block rounded-2xl border p-4 transition hover:bg-white ${
-                      unread
-                        ? "border-brand/40 bg-white shadow-[0_0_0_1px_rgba(82,102,234,0.18),0_18px_45px_-26px_rgba(82,102,234,0.4)]"
-                        : "border-black/10 bg-surface/80 hover:border-black/30 dark:border-white/10 dark:hover:border-white/30"
-                    }`}
-                  >
-                    <div className="flex items-start gap-3">
-                      <div className="min-w-0 flex-1">
-                        <div className="flex flex-wrap items-center gap-2">
-                          <span
-                            className={`inline-flex items-center gap-1 rounded-full px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] uppercase ring-1 ${meta.tone}`}
-                          >
-                            {meta.icon}
-                            {meta.label}
-                          </span>
-                          {unread ? (
-                            <span className="inline-flex items-center gap-1 rounded-full bg-brand px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] text-white uppercase">
-                              {t("badges.newReply")}
+                return (
+                  <li key={r.id}>
+                    <Link
+                      href={`/my-feedback/${r.id}`}
+                      className={`block rounded-2xl border p-4 transition hover:bg-white ${
+                        unread
+                          ? "border-brand/40 bg-white shadow-[0_0_0_1px_rgba(82,102,234,0.18),0_18px_45px_-26px_rgba(82,102,234,0.4)]"
+                          : "border-black/10 bg-surface/80 hover:border-black/30 dark:border-white/10 dark:hover:border-white/30"
+                      }`}
+                    >
+                      <div className="flex items-start gap-3">
+                        <div className="min-w-0 flex-1">
+                          <div className="flex flex-wrap items-center gap-2">
+                            <span
+                              className={`inline-flex items-center gap-1 rounded-full px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] uppercase ring-1 ${meta.tone}`}
+                            >
+                              {meta.icon}
+                              {meta.label}
                             </span>
-                          ) : replied ? (
-                            <span className="inline-flex items-center gap-1 rounded-full bg-chip-success-bg px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] text-chip-success-fg uppercase ring-1 ring-chip-success-fg/20">
-                              {t("badges.replied")}
+                            {unread ? (
+                              <span className="inline-flex items-center gap-1 rounded-full bg-brand px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] text-white uppercase">
+                                {t("badges.newReply")}
+                              </span>
+                            ) : replied ? (
+                              <span className="inline-flex items-center gap-1 rounded-full bg-chip-success-bg px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] text-chip-success-fg uppercase ring-1 ring-chip-success-fg/20">
+                                {t("badges.replied")}
+                              </span>
+                            ) : waiting ? (
+                              <span className="inline-flex items-center gap-1 rounded-full bg-chip-warning-bg px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] text-chip-warning-fg uppercase ring-1 ring-chip-warning-fg/20">
+                                {t("badges.waiting")}
+                              </span>
+                            ) : null}
+                            <span className="ml-auto font-mono text-[10px] tracking-[0.12em] text-muted-4 uppercase">
+                              {new Date(r.createdAt).toLocaleDateString()}
                             </span>
-                          ) : waiting ? (
-                            <span className="inline-flex items-center gap-1 rounded-full bg-chip-warning-bg px-2 py-0.5 font-mono text-[10px] tracking-[0.12em] text-chip-warning-fg uppercase ring-1 ring-chip-warning-fg/20">
-                              {t("badges.waiting")}
-                            </span>
-                          ) : null}
-                          <span className="ml-auto font-mono text-[10px] tracking-[0.12em] text-muted-4 uppercase">
-                            {new Date(r.createdAt).toLocaleDateString()}
-                          </span>
-                        </div>
-
-                        <p className="mt-2 line-clamp-2 text-sm leading-6 text-foreground">
-                          {r.message}
-                        </p>
-
-                        {lastAdminBody ? (
-                          <div className="mt-3 flex items-start gap-2 rounded-xl border border-emerald-200/60 bg-emerald-50/40 p-2.5 dark:border-emerald-800/40 dark:bg-emerald-950/30">
-                            <div className="grid size-6 shrink-0 place-items-center rounded-full bg-emerald-600 font-mono text-[10px] font-semibold text-white">
-                              H
-                            </div>
-                            <div className="min-w-0 flex-1">
-                              <div className="flex items-center gap-1.5 font-mono text-[10px] tracking-[0.12em] text-emerald-900/70 uppercase">
-                                {t("replyAuthor")}
-                                {lastAdminAt ? (
-                                  <span>
-                                    ·{" "}
-                                    {new Date(lastAdminAt).toLocaleDateString(
-                                      undefined,
-                                      { month: "short", day: "numeric" },
-                                    )}
-                                  </span>
-                                ) : null}
-                              </div>
-                              <p className="mt-0.5 line-clamp-2 text-sm leading-6 text-emerald-900 dark:text-emerald-300">
-                                {lastAdminBody}
-                              </p>
-                            </div>
                           </div>
-                        ) : null}
 
-                        <p className="mt-2 text-xs text-muted-3">
-                          {replyCount > 0
-                            ? t("replyCount", { count: replyCount })
-                            : t("noReplies")}
-                        </p>
+                          <p className="mt-2 line-clamp-2 text-sm leading-6 text-foreground">
+                            {r.message}
+                          </p>
+
+                          {lastAdminBody ? (
+                            <div className="mt-3 flex items-start gap-2 rounded-xl border border-emerald-200/60 bg-emerald-50/40 p-2.5 dark:border-emerald-800/40 dark:bg-emerald-950/30">
+                              <div className="grid size-6 shrink-0 place-items-center rounded-full bg-emerald-600 font-mono text-[10px] font-semibold text-white">
+                                H
+                              </div>
+                              <div className="min-w-0 flex-1">
+                                <div className="flex items-center gap-1.5 font-mono text-[10px] tracking-[0.12em] text-emerald-900/70 uppercase">
+                                  {t("replyAuthor")}
+                                  {lastAdminAt ? (
+                                    <span>
+                                      ·{" "}
+                                      {new Date(lastAdminAt).toLocaleDateString(
+                                        undefined,
+                                        { month: "short", day: "numeric" },
+                                      )}
+                                    </span>
+                                  ) : null}
+                                </div>
+                                <p className="mt-0.5 line-clamp-2 text-sm leading-6 text-emerald-900 dark:text-emerald-300">
+                                  {lastAdminBody}
+                                </p>
+                              </div>
+                            </div>
+                          ) : null}
+
+                          <p className="mt-2 text-xs text-muted-3">
+                            {replyCount > 0
+                              ? t("replyCount", { count: replyCount })
+                              : t("noReplies")}
+                          </p>
+                        </div>
                       </div>
-                    </div>
-                  </Link>
-                </li>
-              );
-            })}
-          </ul>
-        )}
-      </section>
-      <SiteFooter />
-    </main>
+                    </Link>
+                  </li>
+                );
+              })}
+            </ul>
+          )}
+        </section>
+        <SiteFooter />
+      </main>
+    </FullAuthProviders>
   );
 }
