+++ src/components/like-button.tsx
@@ -1,154 +1,73 @@
 "use client";
 
-import { useEffect, useRef, useState, useTransition } from "react";
+import { useEffect, useState } from "react";
 
-import { useAuth, useClerk } from "@clerk/nextjs";
 import { Heart } from "lucide-react";
 import { useLocale } from "next-intl";
 
 import { formatLocalizedNumber } from "@/lib/format-number";
 import { loadPetMetrics } from "@/lib/pet-metrics-client";
 
-import { useHeaderState } from "@/components/header-state-provider";
+import { useAuthIntent } from "@/components/auth-intent";
 import { Button } from "@/components/ui/button";
 
 type LikeButtonProps = {
   slug: string;
 };
 
+type LikeButtonComponent = React.ComponentType<LikeButtonProps>;
+
 export function LikeButton({ slug }: LikeButtonProps) {
+  const { authActive, requestAuth } = useAuthIntent();
   const [count, setCount] = useState<number | null>(null);
-  const [liked, setLiked] = useState(false);
-  const [likeStateLoading, setLikeStateLoading] = useState(false);
-  const [pending, start] = useTransition();
+  const [loading, setLoading] = useState(true);
+  const [AuthLikeButton, setAuthLikeButton] =
+    useState<LikeButtonComponent | null>(null);
   const locale = useLocale();
-  const { isLoaded, isSignedIn } = useAuth();
-  const clerk = useClerk();
-  const loadVersionRef = useRef(0);
-  const mutationVersionRef = useRef(0);
-  const { refresh } = useHeaderState();
 
   useEffect(() => {
-    const loadVersion = loadVersionRef.current + 1;
-    loadVersionRef.current = loadVersion;
+    if (!authActive || AuthLikeButton) return;
+    let cancelled = false;
+    void import("@/components/like-button-auth").then((mod) => {
+      if (!cancelled) setAuthLikeButton(() => mod.LikeButton);
+    });
+    return () => {
+      cancelled = true;
+    };
+  }, [AuthLikeButton, authActive]);
 
-    if (!isLoaded) {
-      setLikeStateLoading(true);
-      return;
-    }
-
-    const mutationVersion = mutationVersionRef.current;
-    const controller = new AbortController();
+  useEffect(() => {
+    if (authActive) return;
     let active = true;
-    setLikeStateLoading(true);
-
-    // Signed-in users hit /like (returns authoritative count + their
-    // liked state). Anon visitors hit /metrics (CDN-cached, just the
-    // count). Both endpoints are safe to ignore on abort/error — the
-    // button stays in its skeleton state.
-    const url = isSignedIn
-      ? `/api/pets/${slug}/like`
-      : `/api/pets/${slug}/metrics`;
-
-    void (
-      isSignedIn
-        ? fetch(url, {
-            signal: controller.signal,
-            headers: { accept: "application/json" },
-          }).then((res) => (res.ok ? res.json() : null))
-        : loadPetMetrics(slug)
-    )
-      .then(
-        (
-          data: { liked?: boolean; count?: number; likeCount?: number } | null,
-        ) => {
-          if (!data) return;
-          if (!active) return;
-          if (loadVersionRef.current !== loadVersion) return;
-          if (mutationVersionRef.current !== mutationVersion) return;
-          if (isSignedIn) {
-            setLiked(Boolean(data.liked));
-            setCount(typeof data.count === "number" ? data.count : 0);
-          } else {
-            setLiked(false);
-            setCount(typeof data.likeCount === "number" ? data.likeCount : 0);
-          }
-        },
-      )
-      .catch((error: unknown) => {
+    setLoading(true);
+    void loadPetMetrics(slug)
+      .then((data) => {
         if (!active) return;
-        if (loadVersionRef.current !== loadVersion) return;
-        if ((error as Error).name !== "AbortError") {
-          setLiked(false);
-        }
+        setCount(typeof data?.likeCount === "number" ? data.likeCount : 0);
+      })
+      .catch(() => {
+        if (active) setCount(0);
       })
       .finally(() => {
-        if (!active) return;
-        if (loadVersionRef.current === loadVersion) {
-          setLikeStateLoading(false);
-        }
+        if (active) setLoading(false);
       });
-
     return () => {
       active = false;
-      controller.abort();
     };
-  }, [isLoaded, isSignedIn, slug]);
+  }, [authActive, slug]);
 
-  function handleClick() {
-    if (!isLoaded || !isSignedIn) {
-      clerk.openSignIn({});
-      return;
-    }
-    if (pending || likeStateLoading) return;
-
-    const nextLiked = !liked;
-    const mutationVersion = mutationVersionRef.current + 1;
-    mutationVersionRef.current = mutationVersion;
-    setLiked(nextLiked);
-    setCount((c) => Math.max(0, (c ?? 0) + (nextLiked ? 1 : -1)));
-
-    start(async () => {
-      try {
-        const res = await fetch(`/api/pets/${slug}/like`, {
-          method: "POST",
-          headers: { "content-type": "application/json" },
-          body: JSON.stringify({ liked: nextLiked }),
-        });
-        if (!res.ok) throw new Error("like failed");
-        const data = (await res.json()) as { liked: boolean; count: number };
-        if (mutationVersionRef.current !== mutationVersion) return;
-        setLiked(data.liked);
-        setCount(data.count);
-        void refresh({ force: true });
-      } catch {
-        if (mutationVersionRef.current !== mutationVersion) return;
-        setLiked(!nextLiked);
-        setCount((c) => Math.max(0, (c ?? 0) + (nextLiked ? -1 : 1)));
-      }
-    });
-  }
-
-  const disabled = pending || !isLoaded || likeStateLoading;
+  if (authActive && AuthLikeButton) return <AuthLikeButton slug={slug} />;
 
   return (
     <Button
       variant="outline"
-      onClick={handleClick}
-      aria-pressed={liked}
-      aria-busy={disabled || undefined}
-      disabled={disabled}
-      className={`h-10 gap-2 rounded-full border px-4 text-sm font-medium transition disabled:opacity-60 ${
-        liked
-          ? "border-rose-300 bg-rose-50 text-rose-700 hover:bg-rose-100"
-          : "border-black/10 bg-surface text-muted-2 hover:border-rose-300 hover:text-rose-700"
-      } dark:bg-rose-950/40 dark:text-rose-300 dark:hover:border-rose-700`}
+      onClick={requestAuth}
+      aria-busy={loading || undefined}
+      className="h-10 gap-2 rounded-full border border-black/10 bg-surface px-4 text-sm font-medium text-muted-2 transition hover:border-rose-300 hover:text-rose-700 disabled:opacity-60 dark:bg-rose-950/40 dark:text-rose-300 dark:hover:border-rose-700"
     >
-      <Heart
-        className={`size-4 transition ${liked ? "fill-rose-500 text-rose-500" : ""}`}
-      />
+      <Heart className="size-4 transition" />
       <span className="font-mono text-xs tracking-[0.08em]">
-        {count === null ? "—" : formatLocalizedNumber(count, locale)}
+        {count === null ? "-" : formatLocalizedNumber(count, locale)}
       </span>
     </Button>
   );
